move forked code to an internal package
