+++ component/otelcol/extension/jaeger_remote_sampling/internal/jaegerremotesampling/README.md
@@ -1,0 +1,5 @@
+# jaegerremotesampling
+
+This package is a fork of
+https://github.com/open-telemetry/opentelemetry-collector-contrib/tree/main/extension/jaegerremotesampling
+which supports string-based sampling sources.
