+++ component/otelcol/extension/jaeger_remote_sampling/internal/jaegerremotesampling/extension_test.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package jaeger_remote_sampling
+package jaegerremotesampling
 
 import (
 	"context"
