+++ component/otelcol/extension/jaeger_remote_sampling/internal/jaegerremotesampling/internal/grpc.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package internal // import "github.com/open-telemetry/opentelemetry-collector-contrib/extension/jaegerremotesampling/internal"
+package internal
 
 import (
 	"context"
