+++ component/otelcol/extension/jaeger_remote_sampling/internal/strategy_store/README.md
@@ -1,0 +1,5 @@
+# strategy_store
+
+This package is a fork of
+https://github.com/jaegertracing/jaeger/blob/v1.38.1/plugin/sampling/strategystore/static/strategy_store.go
+to support a string-based strategy store.
